% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD-MDS.R
\name{SVDMDS}
\alias{SVDMDS}
\title{Construction of a SVD-MDS representation}
\usage{
SVDMDS(dist, k = 2, metric = "euclidean", max_it = 6 * 10^6,
  stress_sd_th = 10^-4, stack_length = 500, verbose = TRUE)
}
\arguments{
\item{dist}{a numeric matrix with all pairwise distances between objects of the representation}

\item{k}{a numeric value specifying the desired number of dimensions in the resulting Reference Map representation}

\item{metric}{a character indicating the distance metric to use ("euclidean" or "manhattan")}

\item{max_it}{a numeric defining the maximal number of steps the algorithm can perform}

\item{stress_sd_th}{a numeric defining the threshold for the standard deviation of Kruskal Stress}

\item{stack_length}{a numeric defining the length of the Kruskal Stress stack (used to compute the standard deviation of the Kruskal Stress)}

\item{verbose}{a boolean enabling the display of debug information at each step of the algorithm}
}
\value{
a list of 3 elements containing the position of the objects ('points' element), the Kruskal Stress ('stress' element), and the Entourage Score ('entourage' element)
}
\description{
This function computes a SVD-MDS representation based on a distance matrix.
}

